requirements.txt
openai
mysql-connector-python
python-dotenv
pandas
fastapi
scikit-learn
transformers
torch
protobuf
tiktoken
sentencepiece
accelerate
hdbscan
streamlit
python-multipart
uvicorn
openpyxl
plotly
wordcloud
streamlit_plotly_events
anywidget